python-ntrip-client/requirements.txt
nrf24==2.0.0
python-dotenv
